requirements.txt
https://github.com/qwertyquerty/pypresence/archive/master.zip
python-dotenv
websockets
